# Conflicts:
#	composer.lock
#	src/web/assets/admintable/dist/js/app.js
#	src/web/assets/admintable/dist/js/app.js.map
#	src/web/assets/cp/dist/cp.js
#	src/web/assets/cp/dist/cp.js.map
#	src/web/assets/cp/dist/css/cp.css
#	src/web/assets/cp/dist/css/cp.css.map
#	src/web/assets/edituser/dist/css/profile.css.map
#	src/web/assets/graphiql/dist/css/graphiql.css.map
#	src/web/assets/installer/dist/css/install.css.map
#	src/web/assets/login/dist/css/login.css.map
#	src/web/assets/plugins/dist/css/PluginManager.css.map
#	src/web/assets/pluginstore/dist/css/app.css
#	src/web/assets/pluginstore/dist/css/app.css.map
#	src/web/assets/routes/dist/css/routes.css.map
#	src/web/assets/systemmessages/dist/css/system_messages.css.map
#	src/web/assets/upgrade/dist/css/UpgradeUtility.css
#	src/web/assets/upgrade/dist/css/UpgradeUtility.css.map

+++ src/elements/actions/DeleteAssets.php
@@ -81,7 +81,7 @@ JS, [static::class]);
 
         try {
             foreach ($query->all() as $asset) {
-                if ($elementsService->canDelete($asset, $user)) {
+                if ($elementsService->canView($asset, $user) && $elementsService->canDelete($asset, $user)) {
                     $elementsService->deleteElement($asset);
                 }
             }
